+++ Modules/Microsoft365DSC/DscResources/MSFT_IntuneDeviceCompliancePolicyAndroidDeviceOwner/MSFT_IntuneDeviceCompliancePolicyAndroidDeviceOwner.psm1
@@ -239,7 +239,7 @@ class IntuneDeviceCompliancePolicyAndroidDeviceOwner : M365DSCResourceBase
                     [System.String[]]$groups = @()
                     foreach ($group in $actionConfiguration.NotificationMessageCCList)
                     {
-                        $groups += (Get-MgGroup -GroupId $group -ErrorAction SilentlyContinue).DisplayName
+                        $groups += (Get-M365DSCIntuneGroup -GroupId $group).DisplayName
                     }
                     $scheduledAction.Add('NotificationMessageCCList', $groups)
                 }