+++ Modules/Microsoft365DSC/DscResources/MSFT_IntuneDeviceCompliancePolicyWindows10/MSFT_IntuneDeviceCompliancePolicyWindows10.psm1
@@ -298,7 +298,7 @@ class IntuneDeviceCompliancePolicyWindows10 : M365DSCResourceBase
                     [System.String[]]$groups = @()
                     foreach ($group in $actionConfiguration.NotificationMessageCCList)
                     {
-                        $groups += (Get-MgGroup -GroupId $group -ErrorAction SilentlyContinue).DisplayName
+                        $groups += (Get-M365DSCIntuneGroup -GroupId $group).DisplayName
                     }
                     $scheduledAction.Add('NotificationMessageCCList', $groups)
                 }