+++ Modules/Microsoft365DSC/DscResources/MSFT_IntuneDeviceCompliancePolicyiOs/MSFT_IntuneDeviceCompliancePolicyiOs.psm1
@@ -213,7 +213,7 @@ class IntuneDeviceCompliancePolicyiOs : M365DSCResourceBase
                     [System.String[]]$groups = @()
                     foreach ($group in $actionConfiguration.NotificationMessageCCList)
                     {
-                        $groups += (Get-MgGroup -GroupId $group -ErrorAction SilentlyContinue).DisplayName
+                        $groups += (Get-M365DSCIntuneGroup -GroupId $group).DisplayName
                     }
                     $scheduledAction.Add('NotificationMessageCCList', $groups)
                 }