+++ Modules/Microsoft365DSC/DscResources/MSFT_IntuneWindowsAutopilotDevicePreparationAutomaticPolicy/MSFT_IntuneWindowsAutopilotDevicePreparationAutomaticPolicy.psm1
@@ -209,7 +209,7 @@ class IntuneWindowsAutopilotDevicePreparationAutomaticPolicy : M365DSCResourceBa
             if ($target.enrollmentTimeDeviceMembershipTargetValidationStatuses.Count -gt 0)
             {
                 $groupId = $target.enrollmentTimeDeviceMembershipTargetValidationStatuses[0].targetId
-                $resolvedGroup = Get-MgGroup -GroupId $groupId -Property "id, displayName"
+                $resolvedGroup = Get-M365DSCIntuneGroup -GroupId $groupId
                 $assignmentResult = $resolvedGroup.displayName
             }
             $results.Add('AssignmentTarget', $assignmentResult)